% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{addIdeasButton}
\alias{addIdeasButton}
\title{IDEAS easy button}
\usage{
addIdeasButton(map, position = "bottomleft", extrainfo = NULL)
}
\arguments{
\item{map}{a leaflet map}

\item{position}{button position}

\item{extrainfo}{additional html to include in the box on click}
}
\description{
Adds an easy button with the IDEAS logo with info on click.

\strong{Important Note}: for the font awesome logo to display correctly either
use the \code{\link[idgeo]{prettyLeaflet}}() function when building your map
(which will automatically add the fontawesome dependencies), or apply
\code{\link[idgeo]{useFontAwesome}}() to your map
}
