% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{addLayerControlTitle}
\alias{addLayerControlTitle}
\title{Add Title to Leaflet Layer Control}
\usage{
addLayerControlTitle(
  map,
  title = "Color Map By:",
  type = c("base", "overlay"),
  style = "text-align:center;font-size:larger;"
)
}
\arguments{
\item{map}{a leaflet map}

\item{title}{string, title of control, default 'Color Map By:'}

\item{type}{layer control type, either 'base' or 'overlay', default 'base'}

\item{style}{string, html style string default \code{text-align:center;font-size:larger;}}
}
\description{
Adds a title to a leaflet layer control
}
