% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{addTitleControl}
\alias{addTitleControl}
\title{Add Title Control to Leaflet Map}
\usage{
addTitleControl(
  map,
  title = NULL,
  subtitle = NULL,
  position = c("topleft", "topright", "bottomleft", "bottomright"),
  layerId = NULL,
  ...
)
}
\arguments{
\item{map}{an existing \code{leaflet} map.}

\item{title}{a character string used as the title. Defaults to \code{NULL}.
Either \code{title} or \code{subtitle} is required.}

\item{subtitle}{a character string used as the subtitle. Defaults to
\code{NULL}. Either \code{title} or \code{subtitle} is required.}

\item{position}{position of control: "topleft", "topright", "bottomleft", or
"bottomright"}

\item{layerId}{the leaflet layerid}

\item{...}{additional HTML options (\code{titlesize}, \code{titlecolor},
etc.) to pass to \code{\link[idgeo]{htmlTitle}}. See \code{?idgeo::htmlTitle}
for full documentation}
}
\description{
This function adds a formatted HTML title to a \code{leaflet}
map. For a standalone HTML title block, see \code{\link[idgeo]{htmlTitle}}
}
