% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{alignControls}
\alias{alignControls}
\title{Align Leaflet Controls}
\usage{
alignControls(map, direction = c("horizontal", "vertical"))
}
\arguments{
\item{map}{a \code{leaflet} map}

\item{direction}{defaults to \code{horizontal}, and is intended to be used
without arguments since a choice of \code{vertical}
(\code{leaflet} default) simply returns the map as is.}
}
\description{
This function allows for side by side \code{leaflet} controls
instead of the default horizontal arrangement. Especially useful with
\code{\link[idgeo]{addTitleControl}} so the "topleft" position isn't below
the zoom controls.
}
