% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{ambulanceIcon}
\alias{ambulanceIcon}
\title{Ambulance Icons}
\usage{
ambulanceIcon(iconColor = "#FFFFFF", ...)
}
\arguments{
\item{iconColor}{default of white}

\item{...}{any additional \code{makeAwesomeIcon} arguments}
}
\description{
This function creates an automatic icon to be used with leaflet maps
that represents an ambulance
Submit \code{leaflet::makeAwesomeIcon} to see the full
documentation for \code{\link[leaflet]{makeAwesomeIcon}}
}
