% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{colorPercent}
\alias{colorPercent}
\title{Percentage Color Mapping}
\usage{
colorPercent(palette = "viridis", domain = c(0, 1), ...)
}
\arguments{
\item{palette}{The colors or color function that values will be mapped to.
Defaults to \code{'viridis'}}

\item{domain}{The possible values that can be mapped. Defaults to
\code{c(0,1)}}

\item{...}{other options passed to \code{\link[leaflet]{colorNumeric}}}
}
\description{
Maps values that are known to be between 0 & 1 to a given
palette using \code{\link[leaflet]{colorNumeric}}
}
