% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_average_distance_centroid.R
\name{generate_average_distance_centroid}
\alias{generate_average_distance_centroid}
\title{Calculate Centroid Distance}
\usage{
generate_average_distance_centroid(x, add_plot = F)
}
\arguments{
\item{x}{a data.frame with columns for \code{x} and \code{y}}

\item{add_plot}{a boolean, draw a plot with the points and hull.}
}
\value{
a list with items for:
\describe{
\item{distances_raw}{a matrix with the distance from the centroid for
each points}
\item{distances_sumz}{a data.frame with summary statistics}
\item{cluster_centroid}{the centroid for the convex hull}
\item{cluster_poly}{the polygon of the convex hull}
}
}
\description{
Draws a polygon around a set of points and calculates the distance and
summary statistics for each point to that centroid.
}
\details{
This function takes a set of points and then draws the convex hull using
the \code{chull} function from grDevices. This hull is then converted to a polygon.
A centroid location is then determined from this polygon and the distance
from the centroid to each point is calculate.
}
