% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{htmlTitle}
\alias{htmlTitle}
\title{Create HTML Title}
\usage{
htmlTitle(
  title = NULL,
  subtitle = NULL,
  titlesize = 22,
  subtitlesize = 16,
  titlebold = T,
  subtitlebold = T,
  titlecolor = "black",
  subtitlecolor = idstyle::wake_gold,
  titlefont = idstyle::wake_sansserif,
  subtitlefont = idstyle::wake_sansserif,
  padding = 10
)
}
\arguments{
\item{title}{a character string used as the title. Defaults to \code{NULL}.
Either \code{title} or \code{subtitle} is required.}

\item{subtitle}{a character string used as the subtitle. Defaults to
\code{NULL}. Either \code{title} or \code{subtitle} is required.}

\item{titlesize}{number indicating size of the title in pixels. Default is
\code{22}.}

\item{subtitlesize}{number indicating size of the subtitle in pixels.
Default is \code{16}.}

\item{titlebold}{logical indicating whether or not the title should be bold.
defaults to \code{TRUE}}

\item{subtitlebold}{logical indicating whether or not the title should be
bold. defaults to \code{TRUE}}

\item{titlecolor}{character string indicating a valid color for the title.
Defaults to black.}

\item{subtitlecolor}{character string indicating a valid color for the
subtitle. Defaults to primary from \code{idstyle} package.}

\item{titlefont}{character string indicating a valid font family for the
title. Defaults to sans serif from \code{idstyle} package.}

\item{subtitlefont}{character string indicating a valid font family for the
subtitle. Defaults to sans serif from \code{idstyle} package.}

\item{padding}{number supplying amount of padding in pixels around the title
block. Default is \code{10}.}
}
\description{
This function creates a formatted HTML block intended for use
as a title block in an HTML widget. For adding a title directly to a
leaflet map, see \code{\link[idgeo]{addTitleControl}}
}
