% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{IDLabelOptions}
\alias{IDLabelOptions}
\title{Leaflet Label Options Defaults for ID}
\usage{
IDLabelOptions(
  textsize = "14px",
  opacity = 0.85,
  fontcolor = idstyle::wake_darkgold,
  fontfamily = idstyle::wake_sansserif,
  additionalstyles = NULL,
  ...
)
}
\arguments{
\item{textsize}{new default is \code{'14px'}}

\item{opacity}{new default is \code{0.85}}

\item{fontcolor}{font color for label, default from \code{idstyle} package}

\item{fontfamily}{font family for label, default from \code{idstyle} package}

\item{additionalstyles}{additional named list of styles (not color or font-family) passed to \code{style} argument of \code{\link[leaflet]{labelOptions}}}

\item{...}{additional arguments passed to \code{\link[leaflet]{labelOptions}}}
}
\description{
Some functions that will override some of the 'leaflet' defaults
See the full documentation for \code{\link[leaflet]{labelOptions}} for details
}
