% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{IDTileOptions}
\alias{IDTileOptions}
\title{Leaflet Tile Options for ID}
\usage{
IDTileOptions(minZoom = 6, maxZoom = 16, ...)
}
\arguments{
\item{minZoom}{min zoom value, default \code{6}}

\item{maxZoom}{max zoom value, default \code{16}}

\item{...}{any other arguments to \code{\link[leaflet]{tileOptions}}}
}
\description{
default tile options ID Research
}
