% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{labelFormatPercent}
\alias{labelFormatPercent}
\title{Percentage Label Format for Leaflet Legend}
\usage{
labelFormatPercent(digits = 0, ...)
}
\arguments{
\item{digits}{the number of digits in the percentage}

\item{...}{other arguments passed to \code{\link[leaflet]{labelFormat}}}
}
\description{
A function that automatically formats a leaflet legend as a
percentage. A special case of \code{\link[leaflet]{labelFormat}} with
\code{suffix = "\%"} and transformation of \code{100*x}
}
