% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{leafletFont}
\alias{leafletFont}
\title{Change Fonts used in Leaflet Map}
\usage{
leafletFont(
  map,
  font = idstyle::wake_sansserif,
  tooltip_color = idstyle::wake_darkgold
)
}
\arguments{
\item{map}{an existing \code{leaflet} map.}

\item{font}{a character string specifying the font-family. Defaults to
sans serif from \code{idstyle} package}

\item{tooltip_color}{character string indicating color of tooltip. default of
primary dark from \code{idstyle} package}
}
\description{
This function changes the default font of a \code{leaflet} map.
}
