% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{legendFixNA}
\alias{legendFixNA}
\title{Fix NA in Leaflet Legend}
\usage{
legendFixNA(map)
}
\arguments{
\item{map}{a \code{leaflet} map}
}
\description{
This function corrects the odd alignment of \code{NA} that
appears in some \code{leaflet} legends. Instead of floating in the right
side of the legend, it will be correctly aligned at the bottom of the
legend.
}
