% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_datasets.R
\docType{data}
\name{local_zips}
\alias{local_zips}
\title{Local ZIP Codes}
\format{
\subsection{\code{local_zips}}{

A data frame with 154 rows and 8 columns:
\describe{
\item{ZIP}{ZIP Code}
\item{Classification}{ZIP Classification: General, P.O. Box, or Unique}
\item{City}{City}
\item{County}{County}
\item{Timezone}{Timezone}
\item{AreaCode}{String of Area Code(s), separated by \code{/}}
\item{Grouping}{ZIP Code Grouping for a higher grain than ZIP, but lower
grain that county, for geographic analysis}
\item{CentralTo}{City that the ZIP is central to, if applicable}

}
}
}
\usage{
data(local_zips)
}
\description{
Dataset of ZIP Codes in Guilford, Alamance, Rockingham, Randolph, Forsyth,
Chatham, Stokes, Davidson, and Caswell Counties.
}
\keyword{datasets}
