% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{multiBaseLegends}
\alias{multiBaseLegends}
\title{Multiple Base Legend Support for Leaflet Maps}
\usage{
multiBaseLegends(map, mapnum = 1, map_id = NULL)
}
\arguments{
\item{map}{a \code{leaflet} map}

\item{mapnum}{optional, identifies the map number when multiple maps are
rendered in the same document. Used in conjunction with \code{map_id},
this fixes any javascript rendering issues that might otherwise
incorrectly impact multiple maps. The actual numbers are irrelevant as
long as each map has a unique number. Defaults to \code{1}}

\item{map_id}{optional, defaults to \code{NULL}. A unique string identifier
of the map (any spaces will be removed) when multiple maps are rendered
in the same document. Used in conjunction with \code{mapnum} to fix
javascript rendering issues}
}
\description{
This function allows for multiple base layer legends that toggle
with selection. Default \code{leaflet} shows all legends regardless of
which base layer is selected.
Note that the \emph{title} of the legend for each layer
(\code{addLegend(title = 'Some Layer')}) must be the same as the \emph{group}
name of the layer (\code{addLayersControl(baseGroups = c('Some Layer'))}).
If there are any \emph{overlay layers}, this function also ensures they
remain in front of the multiple \emph{base layers} when toggling.
}
