% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_datasets.R
\docType{data}
\name{nc_counties}
\alias{nc_counties}
\title{North Carolina Counties}
\format{
\subsection{\code{nc_counties}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{county}{County Name}
\item{county_code}{County FIPS Code (3-digits, no State ID)}
\item{GEOID}{State-County FIPS Code}
}
}
}
\usage{
data(nc_counties)
}
\description{
Dataset of NC Counties & FIPS
}
\keyword{datasets}
