% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{prettyLeaflet}
\alias{prettyLeaflet}
\title{Pretty Leaflet Maps}
\usage{
prettyLeaflet(
  ...,
  tiles = T,
  tilesminimal = T,
  align = T,
  legendNA = T,
  font = T,
  fullscreen = T,
  multibase = F,
  idbtn = T,
  btnposition = "bottomleft",
  btnextrainfo = NULL,
  mini = F,
  miniposition = "topright"
)
}
\arguments{
\item{...}{any parameters supplied to \code{\link[leaflet]{leaflet}}}

\item{tiles}{logical, default \code{TRUE}, whether or not to use
\code{\link[idgeo]{addIDProviderTiles}}, the Wake ID default map layer.}

\item{tilesminimal}{whether to use the minimal or more detailed Wake ID default
map tiles. Ignored if \code{tiles} is \code{FALSE}}

\item{align}{logical, default \code{TRUE}, whether or not to use
\code{\link[idgeo]{alignControls}}, horizontal alignment of multiple
\code{leaflet} controls.}

\item{legendNA}{logical, default \code{TRUE}, whether or not to use
\code{\link[idgeo]{legendFixNA}}, which corrects odd placement of
\code{NA} values in the legend.}

\item{font}{logical, default \code{TRUE}, whether or not to use
\code{\link[idgeo]{leafletFont}}, the wake default font face through
the map.}

\item{fullscreen}{logical, default \code{TRUE}, whether or not to use
\code{\link[leaflet.extras]{addFullscreenControl}}, a fullscreen control
button}

\item{multibase}{logical, default \code{FALSE}, whether or not to use
\code{\link[idgeo]{multiBaseLegends}}, support for multiple base layer
legends that toggle appropriately. If you are not careful with the
naming of legends, this may make your legends disappear, and is thus not
on by default.}

\item{idbtn}{default \code{TRUE} whether or not to add easy button with IDEAS logo & info to the map}

\item{btnposition}{position of IDEAS button. ignored
if \code{idbtn} is set to \code{FALSE}}

\item{btnextrainfo}{extra html to include in the button popup on click. ignored
if \code{idbtn} is set to \code{FALSE}}

\item{mini}{whether to add to mini map via \code{\link[leaflet]{addMiniMap}}}

\item{miniposition}{position of mini map. ignored
if \code{mini} is set to \code{FALSE}}
}
\description{
This function can be used in place of
\code{\link[leaflet]{leaflet}} and will apply many helper and style
functions for quick consistent maps. Each added option can be removed
by setting the argument to \code{FALSE}
}
