% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_pts_in_polygon.R
\name{put_pts_in_polygon}
\alias{put_pts_in_polygon}
\title{Put x/y points into polygons}
\usage{
put_pts_in_polygon(df_pts, polygon, keep_geometry = FALSE)
}
\arguments{
\item{df_pts}{a data frame with columns 'x', 'y', or roughly 'lat', 'lon'}

\item{polygon}{an sf object that contains polygons in which to assign the
x/y values to}

\item{keep_geometry}{default \code{FALSE} does not return geometry from the
polygon data.}
}
\description{
identifies which polygon the given x/y coordinates fall in.
The function will convert the data frame with the coordinates into an sf
object so that a \code{st_within} join can be performed. Returned object will
have the same number of rows as \code{df_pts}, all original columns from
\code{df_pts}, along with any columns from \code{polygon}
}
