% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{quickPal}
\alias{quickPal}
\title{Quick Leaflet Palette}
\usage{
quickPal(x, palette = NULL, percent = FALSE, reverse = FALSE, domain = NULL)
}
\arguments{
\item{x}{a vector}

\item{palette}{optional, a string specifying a color palette function or a
vector of colors; default is 'viridis' for numeric or percent
and Okabe-Ito for character or factor (note: Okabe-Ito only supports up to 9
distinct factor levels)}

\item{percent}{default \code{NULL}}

\item{reverse}{whether to reverse the \code{palette}}

\item{domain}{domain used for the \code{palette}. the default \code{domain = NULL} will
use all observations in \code{x} as the range, except when \code{percent = TRUE} in
which case the domain defaults to \code{c(0,1)}}
}
\description{
This function will generate a palette by either
\code{\link[leaflet]{colorNumeric}} or \code{\link[leaflet]{colorFactor}}
for numeric and character/factor vectors respectively. When \code{percent = TRUE}
\code{\link[idgeo]{colorPercent}} will be used instead
}
