% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_datasets.R
\docType{data}
\name{sf_nc_counties}
\alias{sf_nc_counties}
\title{North Carolina Counties with Geometry}
\format{
\subsection{\code{sf_nc_counties}}{

A sf data frame with 100 rows and 18 columns:
\describe{
\item{NAME}{County Name}
\item{COUNTYFP}{County FIPS Code (3-digits, no State ID)}
\item{GEOID}{State-County FIPS Code}
\item{geometry}{County geometry}
}
}
}
\usage{
data(sf_nc_counties)
}
\description{
Dataset of NC Counties with Geometry generated from
\code{tigris::counties('NC')}. See \code{?tigris::counties} for info on
all columns generated. Only partially documented below
}
\keyword{datasets}
