% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_datasets.R
\docType{data}
\name{sf_nc_tracts}
\alias{sf_nc_tracts}
\title{North Carolina Tracts with Geometry}
\format{
\subsection{\code{sf_nc_tracts}}{

A sf data frame with 2,672 rows and 13 columns:
\describe{
\item{GEOID}{Tract FIPS Code}
\item{geometry}{Tract geometry}
}
}
}
\usage{
data(sf_nc_tracts)
}
\description{
Dataset of NC Tracts with Geometry generated from
\code{tigris::tracts('NC')}. See \code{?tigris::tracts} for
info on all columns generated. Only partially documented below
}
\keyword{datasets}
