% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_datasets.R
\docType{data}
\name{sf_nc_zips}
\alias{sf_nc_zips}
\title{North Carolina ZCTAS with Geometry}
\format{
\subsection{\code{sf_nc_zips}}{

A sf data frame with 808 rows and 13 columns:
\describe{
\item{ZIP}{ZIP Code}
\item{geometry}{ZCTA geometry}
}
}
}
\usage{
data(sf_nc_zips)
}
\description{
Dataset of NC ZCTAS with Geometry generated from
\code{tigris::zctas('NC')}. See \code{?tigris::zctas} for
info on all columns generated. Only partially documented below
}
\keyword{datasets}
