% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{useFontAwesome}
\alias{useFontAwesome}
\title{Use Font Awesome}
\usage{
useFontAwesome(map)
}
\arguments{
\item{map}{a leaflet map}
}
\description{
Applies header info for using fontawesome icons (newer version
than what comes with leaflet by default)
}
