% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_controls.R
\name{useWakeCss}
\alias{useWakeCss}
\title{Use Wake CSS}
\usage{
useWakeCss(map)
}
\arguments{
\item{map}{a leaflet map}
}
\description{
Applies header info for using wake css within a leaflet
map. This is only needed for free standing leaflet maps.
\strong{This is not needed if you are including a map within a quarto
\code{idreport} from the \code{idstyle} package (applied within automatically)}
}
